Moved ValueMapping logic and tests to separate files
